Back Button & Quit Game Implementation
I tried multiple ways to implement the back button and I found the best way to do it was to Create a GameObject within the Back Button function, and use the specific window game object to set it to false when the user clicks on the button. The Quit button should world although I did some research and saw that you can't see it in the editor. I am a bit confused on the transform part of the canvas so I would like to go over that again.