In Game Menu Implementation
This push will allow the user to press the escape key to open the in game menu, allow them to click on all the options with functionality, and also having the user go back to the main menu screen.

I'm not sure if this is the most optimal way to do it so please let me know what you think of it.